Fix another lost matches bug
In the process of checking content conditions, there are cases where we
want to stop matching early because we already know the outcome.

Case 1 is when negative condition (e.g. nomatch) matches, and here we
instantly know that the file condition doesn't match.

Case 2 is when all conditions are satisfied, that is when
- there are no negative conditions (which are never satisfied and need
  the whole content)
- no condition is reporting target (otherwise we need whole content to
  get more potential match reports)
- all positive conditions are satistifid (for example, all match
  conditions were matched at least once)

The applier logic relies on Vec::retain to exclude file conditions with
known outcome from consideration and checks. It was OK to drop Case 1
conditions, but when Case 2 conditions were dropped, it also excluded
then from finalization phase, where global match flags were set and
file/line matches were propagated to the upper level.

Fix this by introducing another container with conditions that need
finalization. Now we drop Case 1 from both container, and Case 2 only
from container with currently checked conditions.