Allow match interruption with size conditions
When deciding for early success, we compare number of satisfied conditions
with total number of conditions under `files`. However since we now have
`size` condition which is checked elsewhere and is effectively always satisfied,
but this is not considered. Fix this by introducing linewise conditions which
may be satisfied. Size is this not linewise.